+++ Unity Utils/Packages/UnityUtils/Runtime/Script Utils/Save System/ISaveData.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace UnityUtils.ScriptUtils.SaveSystem
+{
+    /// <summary>
+    /// Implimenting this interface means you must have "[System.Serializable]" above your class name
+    /// </summary>
+    public interface ISaveData { }
+}